Add a `clone` command
Since the new command doesn't really need the `hub.upstream` configuration
variable, a new mechanism to require it only for the other commands had to
be introduced.

Also the configuration `hub.pullurltype` has been renamed to `hub.urltype`
to make more clear it's not only used when pulling (the new command uses
it too).