Make build stop on errors
Both the bash-completion and the man page are generated through pipes and
because of that, the files are created even when there is an error building
those files, making a second `make` not trying to build it again.

Also the generate-bash-completion didn't quit with a proper status code
when there was a problem with the git-hub command, making the build
continue as normal.