Authenticate using OAuth
For this a new command is introduced, git hub setup, to ask for the
authorization token. This way we don't need to store the user's password
anymore. The `hub.password` variable is not used anymore and if is present
a warning is issued to tell the user to remove it.