Make git_config() return None if no config was found
Also make the default not throw and accept None as a default. If you
want git_config() to throw if a value is not present, just use
NO_DEFAULT as default.

This should fix Config.check() too, because missing values were expected
to be None but where '' instead, so the check always succeeded.