Add --create-branch option for pull new/attach
This new option creates a remote branch with a different name than the
head and uses it for the actual pull request. This is very useful if you
have a hot-fix in master and you want to create a pull request without
bothering to create an actual local branch for it (but you still don't
want to use master as the branch name for the pull request).