Add --continue, --skip and --abort to pull rebase
This turned out to be very difficult to implement, and it still has some
rough edges and stuff to improve. The RebaseCmd inside the PullCmd became
too big so having it so nested made the code very hard to read, so it was
moved to the top-level indentation and then added to the PullCmd after it
is completely defined.

A file storing the state of the current rebase is stored in
.git/HUB_PULL_REBASING.