Fix #7: Use the tracking branch as default base branch
When creating a pull request (pull new) or converting an issue into
a pull request (pull attach), if the base branch is not specified, try
first to get the branch name from the remote tracking information
(branch.<ref>.merge configuration). If the branch is not tracking
a remote branch, fallback to the hub.pullbase configuration as before.