Fix #4: Only try to remove a branch when it exists
When running the git hub pull --abort action, the temporary branch used for
the rebase is removed and if there is an error the situation is explained,
but there is no way to tell, from the git command exit status, if the error
was because the branch didn't exist or something else.

To fix this we just first check if the branch actually exists, and only
try to remove it if it does.