Fix #24: correctly unstash and remove temporary stash
Before starting a pull rebase, user modifications are stashed.
A specially formatted message (that hopefully the user will never use by
himself) is used to identify the automatically stashed changes to be
unstashed.

The regular expression used to match the message is incorrect, becasue
by default Python matches the whole line as if '^<regex>$' were written.

A leading ".*" is added to make it match the line with the stash prefix
(stash@{0} On ...).