Fix #52: Use debian version when not in a git repo
When building the man, we need to get the current program version. As
long as the precise git version can be used, is the best way to go, but
if a tarball was downloaded instead of getting the program from the git
repository, we fall back to the debian version.