pull rebase: Force the rebase by default
This commit adds a new config variable `hub.forcerebase` which defaults
to "true". The effect of using this is passing `--force` to the git
rebase command, thus forcing the re-committing of all the commits in the
pull request, which rewrites the Committer and CommitDate metadata, now
pointing to the person doing the rebasing and to the date the rebase was
performed.

This helps keeping the history on who and when incorporated changes to
the blessed repository, keeping the authoring information intact and
without resorting to merges.

As a side effect, all the hashes of the commits in the pull request will
change, but that's expected if the pull request was properly based in
the base branch, so it shouldn't be completely unexpected.

This fixes #59.