Replace ~ with - in the Debian version
We are switching to use Semantic Versioning 2.0.0. For this, we need to
encode the release candidate as vX.Y.Z-rcR instead of vX.YrcR. The
previous versioning scheme was also broken in terms of Debian
versioning, because vX.Y is considered earlier than vX.Yrc1 by Debian).

Unfortunately, Debian versioning scheme is not compatible with SemVer,
the `-` separator is used to specify a Debian revision and we don't have
one because this is a native Debian package, so `-` is not allowed in
this case (and even if we wanted to use it vX.Y is still considered
older than vX.Y-rcZ by Debian).

For this reason, we will use the `~` separator for Debian version, which
works as expected, but we need to keep using the `-` version to be
SemVer compatible.