Make the Python interpreter easier to configure
Python is a bit dependant on the installation/distribution. For example
in older Ubuntus, python2.7 is not the default. Other distributions
don't use the version in the binary. To overcome these problems there
was an incomplete support to install the command to use a particular
python binary, but this didn't work if the "building" environment wasn't
using python 2.7 as the default Python interpreter.

This commit is intended to fix this, by using `make PYTHON=mypython`.