Use fpm to create the debian package
Using the regular debian infrastructure to create the packages is very
overwhelming. Also, one debian directory should be used for each
debian/ubuntu distribution, which is way out of the scope of this tool.

If at some point a serious debian/ubuntu packages wants to make proper
packages, he/she can do so in an external repository, this is how
normally debian packaging works.

To do this some stuff about version numbers and released had to be
adjusted too.