Fix issue/pull request help messages
The variable used to change the name of the object being described was
completely wrong, because it was a class object so it never varied. When
possible, now `cls` is used instead, and when is not the object name is
just specified manually.

A few help typos are also fixed.