pull rebase: Add --stash-(include-untracked|all) options
To address #22 in a simple way, the new options --stash-all and
--stash-include-untracked were added to the pull rebase command, which
translate directly into --all and --include-untracked when invoking the
git stash command.

Previously --all was always used, and caused a lot of trouble with
repositories storing big data files that were outside git control. Now
the default is changed to do a plain git stash, at the expense of
slightly bigger probabilities of getting some conflict in the checkout
of the pull request when rebasing. So to keep using the safe option, now
is necessary to pass --stash-all explicitly.