Use GitHub media type v3
GitHub is changing the default media type used soon. The changes
compared to the current default (beta) are minimum and this tool
shouldn't be affected. Just to make sure we are using a fixed version
(which now is the latest), we explicitly ask for the
"application/vnd.github.v3+json" media type now.

For more details see:
https://developer.github.com/changes/2014-01-07-upcoming-change-to-default-media-type/