setup: Accept e-mail as username too
This is not a trivial change, as GitHub doesn't support this for the
API, so we need to search for the possible username first. Only if one
(and only one) username is found the results of the search is used as
the username. The search is only triggered if there is a `@` symbol in
the username.

Fixes #102.