setup: Ask the user for another note if the current is taken
The GitHub API changed, and it doesn't return the OAuth token anymore.
If a previous note exists, now the user is asked for a new name to
create a new token. If no new name is specified, the setup command is
cancelled. A tip is included to the user in case they want to regenerate
the token of the offered note.