Check repo owner against username when cloning
When cloning a repository, if the repository owner was also given, then
check if the username matches the repository owner. If so, simply treat
it as if the owner name was not given at all in the clone command. This
prevents an untruthful "forking repo" message and avoids error after
cloning.

Fixes #160.