clone: Get smarter at interpretting URLs
Some very half baked support to use full GitHub clone URLs was added to
clone at some point, but this support is insuficient. It completely
ignores the URL itself so it will always clone using the current
`hub.urltype` scheme (which defaults to `ssh_url` and is probably too
obscure to ask the user to manually change it).

This commit makes the clone command to parse the URLs more intelligently
and set the `hub.urltype` config option as appropriate (it seems more
reasonable to think the user will like to use that instead, not only for
the cloning but also for other future operations, like pushing).

Fixes #149.