Fix pull show threading
This is almost a rewrite of pull show's logic. This was a bit broken due
to misunderstanding of the GitHub API results (which are basically
completely undocumented, so it is still uncertain if this commit is
getting it right, but testing seems to indicate it's matching what the
GitHub web shows).

The formatting is also improved to match other changes in the GitHub
API:

* Diff hunks are limited to the last 5 lines
* Comments on a diff hunk are indented to make the threading more visual
* Comments bodies and URLs are rearranged to make them more readable

This also fixes #175.