Only remove lines if they start with "# "
In the documentation is stated that only lines starting with "# " are
considered comments (note the space after #), but this is not what
clean_message() currently does.

Also, even when is a bit vague in the documentation, this commit only
consider as comments lines beginning strictly with # (if it has leading
spaces is not considered a comment anymore).

We are still removing lines that consist on only a # (even when is not
stated specifically in the documentation, I think is what everybody
expects, and otherwise the very help message we show breaks everything).

This commits make the program behaves as the documentation specifies.

Fixes #186.