Fix cloning when a fork with the same name exists
When cloning a repository `hub clone` but the current user already has
a repository with the same name, that repository is used even when it
not a real fork of the repository passed by argument.

This commit searches through the upstream repository forks instead of
the user's repositories to see if a fork already exists. If it doesn't,
a new fork will be created, with a name chosen by GitHub (it will add
a suffix `-1` at the time of writing this). Unfortunately there seem to
be no way to customize that.

A new message is added to print the actual location of the new fork, in
case the name is different from upstream.