clone: Always try to fork
There is no efficient way to search for a fork using the GitHub API, so
the only efficient option is to just try to fork. If a fork already
exists, GitHub will return the existing fork instead of creating a new
one (not documented behavior so there is a small risk on relying on
this).

Fixes #193.