rebase: Merge --ff-only if the PR is a merge
If the HEAD of the PR is a merge commit (has more than 1 parent), ask
the user if they would like to do a `merge --ff-only` instead, as
normally rebasing a merge commit results in horrors and nightmares.

This feature should be considered experimental, in the future this might
be done by default without asking or an option might be added to force
this behaviour beforehand to avoid annoying interactive questions.

Fixes #183.