rebase: Don't fail fatally if we can't push to PR
The update to the PR branch to make sure GitHub detects the merge can
fail if the user rebasing the PR doesn't have permissions over the repo
from where the PR was created (for example is another user's fork).

Because of this, the failure to push to the PR branch should be
non-fatal.