Validate received URLs before cloning
A malicious GitHub server could send an URL with the form
`ext::<command>` and that would run arbitrary code where the git-hub
command is ran. To avoid surprises, a simple heuristic is used to spot
fishy URLs (including any `<transport>::` URL or URLs that don't match
the urltype requested).

This should fix most of #197.