Revert pushing rebased commits also to the PR branch
This commit revert both d5c31317561b7468da4962ed3bda7805cca03411 and
aac3d985db40a39b457a9e4dbe280c52d4f9d2f1.

This feature is of little use as it only works when the user rebasing
the PR can push to the repository where the PR branch lives (which is
almost never the case) and can cause problems with integration of
statuses and CI, as it seems like builds with old hashes are triggered.

We should start using the GitHub API to rebase soon instead anyway.