Retry if a freshly done fork can't by fetched
We retry with some sort of more human exponential backoff if we can't
clone or fetch a repo that we just forked, because forking is
asynchronous in GitHub. The wait time is of about 5m which is the
maximum specified by GitHub, and after that they suggest contacting
support, so we do the same:
https://developer.github.com/v3/repos/forks/#create-a-fork

Fixes #214.