Use consistent names for upstream and fork
Historically the names of the remotes of the upstream repo and fork changed
based on the use of --triangular, because for the originally cloned repo
`origin` was always used as the remote name.

With the change to enable --triangular by default, users pushing explicitly
to origin that are used to origin being the fork (which is how it works
when --no-triangular is used) could get nasty surprises by now finding
themselves pushing to upstream instead.

To avoid this, the remotes are now named consistently, no matter if
--triangular was used or not. By default upstream will always be named
upstream and fork will be named fork, so at least the safe intersection of
both can still be used. Changing the variable name will also avoid
accidental pushes as pushing to origin will be an error unless further
configuration was performed.

For people wanting to stick with --no-triangular by default and want to get
the old naming of remotes, is enough to do:

  git config --global hub.triangular false
  git config --global hub.forkremote origin

For the people that want to use the new --triangular default but want to
get the old naming of remotes, is enough to do:

  git config --global hub.upstreamremote origin

Also new command-line options to change this when git-hub clon'ing are
added: -U/--upstreamremote and -F/--forkremote and the old -r/--remote
option was removed because its meaning depended also on --triangular, which
is confusing and error prone.