Fix pull attach when the PR is created from a fork
GitHub added a parameter to the API to specify if maintainers can modify
the PRs, which defaults to true. For this to be possible, the
maintainers of the original repo need push access over the fork, which
not always happen, and when it doesn't the PR code attach API call
fails.

This commit always passes `maintainer_can_modify=false` to make sure the
call will always succeed. This might not be ideal for cases where the
user really want to allow maintainers to modify the PR, so in the future
it might be needed to add some option to allow this from the CLI if
users request it.

More info:
* https://developer.github.com/v3/pulls/#create-a-pull-request
* https://help.github.com/articles/allowing-changes-to-a-pull-request-branch-created-from-a-fork/

Fixes #218.