Update setup command to use a pre-existing PAT
GitHub doesn't support other means of authentication other than PAT
(Personal Access Token) for a while now, so there is no way to create a
PAT ourselves by using basic auth as in the past.

This completely broke the setup command, which now fails
unconditionally, creating a very very bad entry point to the tool (and
probably leaving most users believing the tool is completely broken).

This commit removes basic authentication completely and adds a way to specify
a PAT explicitly via the setup command. It also adds some help on how to create
the PAT.

Fixes #288 (and it is somehow related to #107).